+++ client/src/components/music/EditAlbumInfo.vue
@@ -1,21 +1,18 @@
 <template>
-  <li class="list-inline-item edit-album-info-button" @click="show"
-    v-if="canEditAlbum">
-    <div class="fw-normal text-body badge rounded-pill border"
-	    >
+  <li class="list-inline-item edit-tags-button" @click="show">
+    <div class="fw-normal text-body badge rounded-pill border">
       <font-awesome-icon class="opacity-50 pe-1" icon="edit" size="sm" />
-      <span class="opacity-75">Edit Info</span>
+      <span class="opacity-75">Edit</span>
     </div>
   </li>
   <Modal
-    title="Edit info"
+    title="Edit Album"
     ref="modal"
     @confirm="setAlbumInfo"
-    confirm-label="Update Info"
+    confirm-label="Update"
   >
     <h3>{{ albumTitle }}</h3>
     <h4 v-if="album.album_artist">by {{ album.album_artist.name }}</h4>
-
     <div class="mt-3">
       <label class="col-sm-3 col-form-label">Label</label>
       <input
@@ -45,51 +42,76 @@
         placeholder="Enter pronunciation"
       />
     </div>
-    <div class="mt-3 form-check">
-    <input
-      v-model="form.is_compilation"
-      type="checkbox"
-      class="form-check-input"
-      id="isCompilation"
-    />
-    <label class="form-check-label" for="isCompilation">
-      Compilation
-    </label>
-</div>
 
+    <div
+      v-for="group in Object.keys(groupedTags)"
+      :key="group"
+      class="w-50 d-inline-block mt-1"
+    >
+      <h5 class="mt-1 mb-1 text-capitalize">{{ group }}</h5>
+      <ul class="list-unstyled">
+        <!-- Create two radio button groups for 'local' and 'rotation' tags,
+          that allow you to select one of the valid tags or 'Neither', a null placeholder value -->
+        <li v-for="tag in groupedTags[group].tags" :key="tag">
+          <input
+            class="me-1"
+            type="radio"
+            :value="tag"
+            v-model="groupedTags[group].selected"
+          />
+          <Tag
+            class="selectable-tag"
+            :class="{ selected: groupedTags[group].selected == tag }"
+            :tag="tag !== null ? tag : 'Neither'"
+            @click="groupedTags[group].selected = tag"
+          />
+        </li>
+      </ul>
+    </div>
   </Modal>
 </template>
 
 <script>
 import Modal from "../ModalDialog.vue";
+import Tag from "./TagBadge.vue";
+
 import { mapStores } from "pinia";
 import { useAlbumsStore } from "@/stores/albums";
-import { useAuthStore } from "@/stores/auth";
+import { allowedTags } from "@/constants";
 
+const tagGroups = ["local", "rotation"];
 
 export default {
-  name: "EditAlbumInfoButton",
+  name: "EditAlbumInfo",
   props: {
-    album: {
-      type: Object,
-      required: true,
+    currentTags: {
+      type: Array,
+      default() {
+        return [];
+      },
     },
+    album: Object,
   },
-  components: { Modal },
+  components: { Modal, Tag },
   data() {
     return {
-      form: {
-        label: this.album.label || "",
-        year: this.album.year || "",
-        pronunciation: this.album.pronunciation || "",
-        is_compilation: this.album.is_compilation || true,
-      },
+      groupedTags: tagGroups.reduce((gts, tagGroup) => {
+        return {
+          ...gts,
+          [tagGroup]: {
+            selected:
+              this.currentTags?.find((tag) => tag.includes(tagGroup)) || null,
+            tags: allowedTags
+              .filter((tag) => tag.includes(tagGroup))
+              .concat(null), //Add a null as the value for a 'neither' option in the checkbox group,
+          },
+        };
+      }, {}),
+     form: this.buildForm(),
     };
   },
   computed: {
     ...mapStores(useAlbumsStore),
-    ...mapStores(useAuthStore),
-
     albumTitle() {
       let title = this.album.title;
 
@@ -99,23 +121,29 @@ export default {
 
       return title;
     },
-    canEditAlbum() {
-      return this.authStore.hasRole("music_director");
-    }
   },
   methods: {
-    show() {
-      this.form = {
+    buildForm() {
+      return {
         label: this.album.label || "",
         year: this.album.year || "",
         pronunciation: this.album.pronunciation || "",
-	is_compilation: this.album.is_compilation || false,
       };
+    },
+    show() {
+      this.form = this.buildForm();
       this.$refs.modal.show();
     },
     setAlbumInfo() {
-      this.albumsStore.updateAlbumInfo({
+      let tags = tagGroups
+        .map((tagGroup) => this.groupedTags[tagGroup].selected)
+        .filter(
+          (tag) =>
+            tag !== null && tag !== undefined && allowedTags.includes(tag)
+        );
+      this.albumsStore.updateAlbum({
         album: this.album,
+        tags,
         label: this.form.label,
         year: this.form.year,
         pronunciation: this.form.pronunciation,
@@ -125,11 +153,16 @@ export default {
   },
 };
 </script>
-
 <style scoped>
-.edit-album-info-button {
+.edit-tags-button {
   cursor: pointer;
 }
+
+.selectable-tag {
+  opacity: 0.6;
+}
+.selectable-tag:hover,
+.selectable-tag.selected {
+  opacity: 1;
+}
 </style>
-
-
